Fix relative and semi-relative
whoops, this is what wanted initially. Note: rel and semi-rel flicker when m64 is playing, which is expected because it doesn't expect data being sent like this. So just switch to Instant when in read-only (because there's no way to tell from tasinput side if it's joystick or m64 data)