Lua: Implement abstract presenter system
The concept of a "Presenter" is introduced by this commit.

A presenter is responsible for holding the d2d1 rt of a lua instance, drawing the lua graphics to a control, and blitting its graphics to arbitrary DCs.

For now, the only presenter is the DComp-based one, which is performant and supports per-pixel transparency but only works on Windows 10+ and isn't currently supported by Wine.

An alternative future presenter, the GDIPresenter, will be introduced to handle the second case.

The user will be provided with a choice between the presenters in the config.